tests: test_app_lifecycle: Add app alert
This patch addresses a TODO in the test to test app alerts.  There is no
need to trigger the alert to test it.  This patch adds a "DEPLOYMENT_LIVE"
alert rule and checks the list of alerts to see that the rule exists.
There is no need to trigger the alert to be able to list the alert; we
are listing the alerts configured for the app.  If there are failures for
this test we may need to wait for the app to begin deployment so that the
spec is evaluated and alerts created; generally this happens very quickly.