tests: test_app_lifecycle: Add app alert and fix deploy (#106)
* tests: test_app_lifecycle:  Fix deploy failure

The app spec has an incorrect `run_command` which will result in the
deploy failing with a "bin/api not found" message.  The `run_command`
entry can be removed entirely as the sample will work with the defaults.

* tests: test_app_lifecycle: Add app alert

This patch addresses a TODO in the test to test app alerts.  There is no
need to trigger the alert to test it.  This patch adds a "DEPLOYMENT_LIVE"
alert rule and checks the list of alerts to see that the rule exists.
There is no need to trigger the alert to be able to list the alert; we
are listing the alerts configured for the app.  If there are failures for
this test we may need to wait for the app to begin deployment so that the
spec is evaluated and alerts created; generally this happens very quickly.

* tests: test_app_lifecycle: Use user email for alert update

The assign_alert_destinations() call requires that the email being added
is a valid email that has access to the app.  This patch queries the
account to get the email address and uses it in the call to assign the
alert destination.